refactor: extraction composant et style :lipstick: :recycle:
