+++ src/features/events/components/eventForm/newEvent.style.css
@@ -1,0 +1,6 @@
+.form-container div {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+
+}
